requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.21
openai>=1.30.0
qrcode==7.4.2
reportlab==4.0.7
whitenoise==6.6.0
gunicorn==21.2.0
Pillow==10.3.0
google-generativeai==0.8.3
dj-database-url==2.1.0
python-socketio==5.8.0
python-engineio==4.7.1

# for deployment of render
gunicorn
psycopg2-binary
dj-database-url
daphne 